modificación de las rutas y controladores con sequelize

+++ server/src/models/Books.js
@@ -1,62 +1,43 @@
 const { DataTypes } = require('sequelize');
 const Sequelize = require('Sequelize');
 
-
 module.exports = (sequelize) => {
-    // defino el modelo
-    sequelize.define('Books', {
-
-
-        id: {
-            type: Sequelize.UUID,
-            defaultValue: Sequelize.UUIDV4,
-            primaryKey: true,
-        }, 
-        title: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-
-
-        authors: {
-           
-            type: DataTypes.STRING,
-            allowNull: false
-
-
-        },
-        publisher: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-       
-
-
-        ISBN: {
-            type: DataTypes.INTEGER,
-            allowNull: false
-
-
-        },
-       
-        imageLinks: {
-            type: DataTypes.STRING,
-           
-        },
-        description: {
-            type: DataTypes.TEXT,
-           
-        },
-        price: {
-            type: DataTypes.FLOAT,
-            allowNull: false
-        }
-
-
-
-    }, {
-        timestamps: false,
-
-    });
+	// defino el modelo
+	sequelize.define('Books', {
+		id: {
+			type: Sequelize.UUID,
+			defaultValue: Sequelize.UUIDV4,
+			primaryKey: true,
+		}, 
+		title: {
+			type: DataTypes.STRING,
+			allowNull: false
+		},
+		authors: {	
+			type: DataTypes.STRING,
+			allowNull: false
+		},
+		publisher: {
+			type: DataTypes.STRING,
+			allowNull: false
+		},
+		ISBN: {
+			type: DataTypes.BIGINT,
+			allowNull: false
+		},
+		imageLinks: {
+			type: DataTypes.STRING,
+		},
+		description: {
+			type: DataTypes.TEXT,
+		},
+		price: {
+			type: DataTypes.FLOAT,
+			allowNull: false
+		}
+	}, 
+	{
+		timestamps: false,
+	});
 };
